Refactor interfaces around a common class
